Added testing for saving/loading rcvd data files

+++ src/gui-helpers/gui-generic-helper.hpp
@@ -53,7 +53,7 @@ public:
     static bool getSaveFilePath(QString *filePath, QString fileTypes = tr("All Files (*)"));
     static bool saveFile(QString filePath, QByteArray data);
     static bool copyFile(QString pathFrom, QString pathTo, bool overwrite = false);
-    static uint32_t getFileSize(QString filePath);
+    static qint64 getFileSize(QString filePath);
     static QByteArray loadFile(QString filePath);
 
     // Read. parse, delete a config INI
